move json wrapper out of contrib
