+++ docs/contrib/wrappers.rst
@@ -2,10 +2,15 @@
 Extra Wrappers
 ==============
 
+.. warning::
+    .. deprecated:: 0.15
+        All classes in this module have been moved or deprecated and
+        will be removed in version 1.0. Check the docs for the status
+        of each class.
+
 .. automodule:: werkzeug.contrib.wrappers
 
 .. autoclass:: JSONRequestMixin
-   :members:
 
 .. autoclass:: ProtobufRequestMixin
    :members:
