+++ tests/contrib/test_wrappers.py
@@ -14,16 +14,6 @@ from werkzeug import routing
 from werkzeug.wrappers import Request, Response
 
 
-def test_json_request_mixin():
-    class MyRequest(wrappers.JSONRequestMixin, Request):
-        pass
-    req = MyRequest.from_values(
-        data=u'{"foä": "bar"}'.encode('utf-8'),
-        content_type='text/json'
-    )
-    assert req.json == {u'foä': 'bar'}
-
-
 def test_reverse_slash_behavior():
     class MyRequest(wrappers.ReverseSlashBehaviorRequestMixin, Request):
         pass
