+++ werkzeug/__init__.py
@@ -22,7 +22,13 @@ from werkzeug._compat import iteritems
 
 __version__ = '0.15.dev'
 
-warnings.filterwarnings("default", category=DeprecationWarning, module="werkzeug")
+warnings.filterwarnings(
+    "default",
+    category=DeprecationWarning,
+    module="werkzeug",
+    append=True,
+)
+del warnings
 
 
 # This import magic raises concerns quite often which is why the implementation
